Updated following peer review.
