+++ src/utilities/utilities.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const moment = require('moment')
+
 module.exports = class Utilities {
   static stripWhitespace (value) {
     if (value) {
@@ -54,5 +56,32 @@ module.exports = class Utilities {
       value = null
     }
     return value
+  }
+
+  // Extracts the day from a date that is in YYYY-MM-DD format
+  static extractDayFromDate (inputDate) {
+    // Return the day of birth, with the leading zero Stripped if there is one
+    return parseInt(inputDate.split('-').pop())
+  }
+
+  // Formats the date object (e.g. { day: 31, month: 1, year: 1970 }) into YYYY-MM-DD format (e.g. 1970-01-31) ready for persistence
+  static formatDateForPersistence (inputDate) {
+    return `${inputDate.year}-${inputDate.month}-${inputDate.day}`
+  }
+
+  // Formats the date object (e.g. { day: 31, month: 1, year: 1970 }) into MMMM YYYY format (e.g. January 1970) for display
+  static formatDateForDisplay (inputDate) {
+    let returnValue
+    if (inputDate && inputDate.month && inputDate.year) {
+      let month = inputDate.month.toString()
+      if (month && month.length === 1) {
+        // Pad with a leading zero if required
+        month = '0' + month
+      }
+      returnValue = moment(`${inputDate.year}-${month}-01`).format('MMMM YYYY')
+    } else {
+      returnValue = 'Unknown date'
+    }
+    return returnValue
   }
 }
